Plane: Quadplane: move Q Assist check into new VTOL assist files.
